Refactor project into package structure
