+++ app/models.py
@@ -1,4 +1,4 @@
-from app import db
+from . import db
 from flask_login import UserMixin
 from datetime import datetime
 from sqlalchemy import func
