+++ app/templates/reports.html
