+++ app/templates/scan.html
