+++ run.py
@@ -1,4 +1,6 @@
-from app import app
+from app import create_app
+
+app = create_app()
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', port=5000, debug=True)
